stupid Python 2.7!!

+++ src/Directory.py
@@ -163,7 +163,7 @@ class Directory:
         return struct.unpack_from("<B",self.inputFile.read(1))[0]
         
     def getAttributeByte(self, index):
-        clusterOffset = self.clusterlist[round(index/self.fatVbr.getClusterSize())]*self.fatVbr.getClusterSize()
+        clusterOffset = self.clusterlist[int(round(index/self.fatVbr.getClusterSize()))]*self.fatVbr.getClusterSize()
         indexOffset = (index % 512) * self.fatVbr.getBytesPrRootDirEntry()
         self.inputFile.seek(self.dataOffset + clusterOffset + indexOffset+11)
         return struct.unpack_from("<B",self.inputFile.read(1))[0]
